requirements.txt
openai
google-auth
google-auth-oauthlib
pytrends
requests
schedule